feat: Validaciones al exportar
